Refactor FormActivity dependency injection.
