+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/form/FormActivityRegistrar.kt
@@ -6,18 +6,20 @@ import com.stripe.android.common.taptoadd.TapToAddHelper
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.paymentelement.confirmation.ConfirmationHandler
 import javax.inject.Inject
+import javax.inject.Singleton
 
 internal interface FormActivityRegistrar {
     fun registerAndBootstrap(
         activityResultCaller: ActivityResultCaller,
         lifecycleOwner: LifecycleOwner,
-        paymentMethodMetadata: PaymentMethodMetadata
     )
 }
 
+@Singleton
 internal class DefaultFormActivityRegistrar @Inject constructor(
     private val confirmationHandler: ConfirmationHandler,
     private val tapToAddHelper: TapToAddHelper,
+    private val paymentMethodMetadata: PaymentMethodMetadata,
 ) : FormActivityRegistrar {
 
     private var isBootstrapped = false
@@ -25,7 +27,6 @@ internal class DefaultFormActivityRegistrar @Inject constructor(
     override fun registerAndBootstrap(
         activityResultCaller: ActivityResultCaller,
         lifecycleOwner: LifecycleOwner,
-        paymentMethodMetadata: PaymentMethodMetadata
     ) {
         confirmationHandler.register(activityResultCaller, lifecycleOwner)
         tapToAddHelper.register(activityResultCaller, lifecycleOwner)
