--- paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/form/FormActivityScope.kt
@@ -1,7 +1,0 @@
-package com.stripe.android.paymentelement.embedded.form
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class FormActivityScope
