--- paymentsheet/src/test/java/com/stripe/android/paymentelement/embedded/form/FakeFormActivityRegistrar.kt
@@ -1,30 +1,0 @@
-package com.stripe.android.paymentelement.embedded.form
-
-import androidx.activity.result.ActivityResultCaller
-import androidx.lifecycle.LifecycleOwner
-import app.cash.turbine.Turbine
-import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
-
-internal class FakeFormActivityRegistrar : FormActivityRegistrar {
-    val registerAndBootstrapTurbine = Turbine<RegisterCall>()
-
-    override fun registerAndBootstrap(
-        activityResultCaller: ActivityResultCaller,
-        lifecycleOwner: LifecycleOwner,
-        paymentMethodMetadata: PaymentMethodMetadata
-    ) {
-        registerAndBootstrapTurbine.add(
-            RegisterCall(
-                activityResultCaller = activityResultCaller,
-                lifecycleOwner = lifecycleOwner,
-                paymentMethodMetadata = paymentMethodMetadata
-            )
-        )
-    }
-
-    data class RegisterCall(
-        val activityResultCaller: ActivityResultCaller,
-        val lifecycleOwner: LifecycleOwner,
-        val paymentMethodMetadata: PaymentMethodMetadata,
-    )
-}
