+++ paymentsheet/src/test/java/com/stripe/android/paymentelement/embedded/form/FakeFormActivityStateHelper.kt
@@ -3,7 +3,6 @@ package com.stripe.android.paymentelement.embedded.form
 import app.cash.turbine.Turbine
 import com.stripe.android.core.strings.ResolvableString
 import com.stripe.android.core.strings.resolvableString
-import com.stripe.android.paymentelement.confirmation.ConfirmationHandler
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.ui.PrimaryButton
 import com.stripe.android.paymentsheet.ui.PrimaryButtonProcessingState
@@ -13,8 +12,6 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
 
 internal class FakeFormActivityStateHelper : FormActivityStateHelper {
-    val updateEnabledTurbine = Turbine<Boolean>()
-    val updateTurbine = Turbine<ConfirmationHandler.State>()
     val selectionTurbine = Turbine<PaymentSelection.Saved?>()
 
     override val state: StateFlow<FormActivityStateHelper.State>
@@ -32,14 +29,6 @@ internal class FakeFormActivityStateHelper : FormActivityStateHelper {
     val resultTurbine = Turbine<FormResult>()
 
     override val result: SharedFlow<FormResult> = MutableSharedFlow<FormResult>()
-
-    override fun updateEnabled(enabled: Boolean) {
-        updateEnabledTurbine.add(enabled)
-    }
-
-    override fun updateConfirmationState(confirmationState: ConfirmationHandler.State) {
-        updateTurbine.add(confirmationState)
-    }
 
     override fun updateMandate(mandateText: ResolvableString?) {
         error("This should never be called!")
@@ -62,8 +51,6 @@ internal class FakeFormActivityStateHelper : FormActivityStateHelper {
     }
 
     fun validate() {
-        updateEnabledTurbine.ensureAllEventsConsumed()
-        updateTurbine.ensureAllEventsConsumed()
         resultTurbine.ensureAllEventsConsumed()
     }
 }
