+++ paymentsheet/src/test/java/com/stripe/android/paymentelement/embedded/form/FormActivityConfirmationHandlerTest.kt
@@ -1,14 +1,10 @@
 package com.stripe.android.paymentelement.embedded.form
 
 import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.testing.TestLifecycleOwner
 import com.google.common.truth.Truth.assertThat
-import com.stripe.android.common.taptoadd.FakeTapToAddHelper
 import com.stripe.android.isInstanceOf
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadataFactory
-import com.stripe.android.model.PaymentIntentFixtures
 import com.stripe.android.model.PaymentMethodFixtures
-import com.stripe.android.paymentelement.confirmation.ConfirmationHandler
 import com.stripe.android.paymentelement.confirmation.FakeConfirmationHandler
 import com.stripe.android.paymentelement.confirmation.PaymentMethodConfirmationOption
 import com.stripe.android.paymentelement.embedded.EmbeddedSelectionHolder
@@ -19,7 +15,6 @@ import com.stripe.android.testing.CoroutineTestRule
 import kotlinx.coroutines.test.runTest
 import org.junit.Rule
 import org.junit.Test
-import org.mockito.Mockito.mock
 
 internal class FormActivityConfirmationHandlerTest {
     @get:Rule
@@ -40,24 +35,6 @@ internal class FormActivityConfirmationHandlerTest {
     }
 
     @Test
-    fun `stateHelper is updated on confirmationHandler state change`() = testScenario {
-        selectionHolder.set(PaymentMethodFixtures.CARD_PAYMENT_SELECTION)
-        formConfirmationHelper.confirm()
-        val args = confirmationHandler.startTurbine.awaitItem()
-        confirmationHandler.state.value = ConfirmationHandler.State.Confirming(args.confirmationOption)
-
-        assertThat(stateHelper.updateTurbine.awaitItem()).isInstanceOf<ConfirmationHandler.State.Confirming>()
-
-        confirmationHandler.state.value = ConfirmationHandler.State.Complete(
-            ConfirmationHandler.Result.Succeeded(
-                intent = PaymentIntentFixtures.PI_SUCCEEDED,
-            )
-        )
-
-        assertThat(stateHelper.updateTurbine.awaitItem()).isInstanceOf<ConfirmationHandler.State.Complete>()
-    }
-
-    @Test
     fun `calls onClickOverride if provided`() = testScenario {
         var didCall = false
         onClickOverrideDelegate.set { didCall = true }
@@ -74,7 +51,6 @@ internal class FormActivityConfirmationHandlerTest {
     private fun testScenario(
         block: suspend Scenario.() -> Unit
     ) = runTest {
-        val formActivityConfirmationHandlerRegistrar = FakeFormActivityRegistrar()
         val confirmationHandler = FakeConfirmationHandler()
         val selectionHolder = EmbeddedSelectionHolder(SavedStateHandle())
         val embeddedState = EmbeddedConfirmationStateFixtures.defaultState()
@@ -89,18 +65,11 @@ internal class FormActivityConfirmationHandlerTest {
             configuration = embeddedState.configuration,
             selectionHolder = selectionHolder,
             stateHelper = stateHelper,
-            lifecycleOwner = TestLifecycleOwner(),
-            activityResultCaller = mock(),
             onClickDelegate = onClickOverrideDelegate,
             eventReporter = FakeEventReporter(),
             customerStateHolder = customerStateHolder,
-            coroutineScope = this,
-            formActivityRegistrar = formActivityConfirmationHandlerRegistrar,
-            tapToAddHelper = FakeTapToAddHelper.noOp(),
+            coroutineScope = backgroundScope,
         )
-
-        assertThat(formActivityConfirmationHandlerRegistrar.registerAndBootstrapTurbine.awaitItem()).isNotNull()
-        assertThat(stateHelper.updateTurbine.awaitItem()).isInstanceOf<ConfirmationHandler.State.Idle>()
 
         Scenario(
             formConfirmationHelper = confirmationHelper,
